wip: rework hal for qemu

+++ components/internal/i4a_hal/include/i4a_hal.h
@@ -10,7 +10,7 @@
 void i4a_hal_init();
 
 /** -- config -- */
-uint32_t hal_get_config_bits();
+uint8_t hal_get_config_bits();
 /** -- config -- */
 
 /** -- spi -- */
